refactor : qna 루트 디렉토리 아래 web 디렉토리 추가
