+++ checklist/미션1.md
@@ -112,7 +112,7 @@ public class MvcConfig extends WebMvcConfigurerAdapter {
 - [X] 컨트롤러 공통 url 묶어주기
   - UserController의 "users"
   - QuestionController의 "questions"
-- [ ] qna 루트 디렉토리 아래 web 디렉토리 추가
+- [X] qna 루트 디렉토리 아래 web 디렉토리 추가
 
 ### 질문하기 기능 구현
