+++ src/main/java/com/codessquad/qna/web/QnaApplication.java
@@ -1,4 +1,4 @@
-package com.codessquad.qna;
+package com.codessquad.qna.web;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
