+++ src/main/java/com/codessquad/qna/web/controller/UserController.java
@@ -1,6 +1,6 @@
-package com.codessquad.qna.controller;
+package com.codessquad.qna.web.controller;
 
-import com.codessquad.qna.model.User;
+import com.codessquad.qna.web.model.User;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
