+++ src/main/java/com/codessquad/qna/web/model/Question.java
@@ -1,4 +1,4 @@
-package com.codessquad.qna.model;
+package com.codessquad.qna.web.model;
 
 public class Question {
     private Integer id;
